Refactor: 기존 domain layer에 vo 패키지와 enums 패키지를 추가
무분별하게 존재하는 도메인과 enum, repository의 구분성을 확실히 함
