+++ server/src/main/java/com/ryc/api/v2/club/domain/ClubRepository.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.Optional;
 
 import com.ryc.api.v2.auth.domain.Admin;
+import com.ryc.api.v2.club.domain.enums.Role;
+import com.ryc.api.v2.club.domain.vo.Club;
 
 public interface ClubRepository {
   Club save(Club club);
