--- server/src/main/java/com/ryc/api/v2/club/domain/Role.java
@@ -1,6 +1,0 @@
-package com.ryc.api.v2.club.domain;
-
-public enum Role {
-  OWNER,
-  MEMBER;
-}
