+++ server/src/main/java/com/ryc/api/v2/club/domain/enums/Role.java
@@ -1,0 +1,6 @@
+package com.ryc.api.v2.club.domain.enums;
+
+public enum Role {
+  OWNER,
+  MEMBER;
+}
