+++ server/src/main/java/com/ryc/api/v2/club/domain/vo/ClubTag.java
@@ -1,4 +1,4 @@
-package com.ryc.api.v2.club.domain;
+package com.ryc.api.v2.club.domain.vo;
 
 import jakarta.persistence.Embeddable;
