+++ server/src/main/java/com/ryc/api/v2/club/infra/ClubRepositoryImpl.java
@@ -5,9 +5,9 @@ import java.util.*;
 import org.springframework.stereotype.Repository;
 
 import com.ryc.api.v2.auth.domain.Admin;
-import com.ryc.api.v2.club.domain.Club;
+import com.ryc.api.v2.club.domain.vo.Club;
 import com.ryc.api.v2.club.domain.ClubRepository;
-import com.ryc.api.v2.club.domain.Role;
+import com.ryc.api.v2.club.domain.enums.Role;
 import com.ryc.api.v2.club.infra.entity.ClubEntity;
 import com.ryc.api.v2.club.infra.entity.RoleEntity;
 import com.ryc.api.v2.club.infra.jpa.ClubJpaRepository;
