+++ server/src/main/java/com/ryc/api/v2/club/infra/entity/RoleEntity.java
@@ -3,7 +3,7 @@ package com.ryc.api.v2.club.infra.entity;
 import jakarta.persistence.*;
 
 import com.ryc.api.v2.auth.infra.entity.AdminEntity;
-import com.ryc.api.v2.club.domain.Role;
+import com.ryc.api.v2.club.domain.enums.Role;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
