+++ server/src/main/java/com/ryc/api/v2/club/infra/mapper/RoleMapper.java
@@ -3,8 +3,8 @@ package com.ryc.api.v2.club.infra.mapper;
 import com.ryc.api.v2.auth.domain.Admin;
 import com.ryc.api.v2.auth.infra.entity.AdminEntity;
 import com.ryc.api.v2.auth.infra.mapper.AdminMapper;
-import com.ryc.api.v2.club.domain.Club;
-import com.ryc.api.v2.club.domain.Role;
+import com.ryc.api.v2.club.domain.vo.Club;
+import com.ryc.api.v2.club.domain.enums.Role;
 import com.ryc.api.v2.club.infra.entity.ClubEntity;
 import com.ryc.api.v2.club.infra.entity.RoleEntity;
