+++ server/src/main/java/com/ryc/api/v2/club/presentation/dto/request/ClubUpdateRequest.java
@@ -3,9 +3,9 @@ package com.ryc.api.v2.club.presentation.dto.request;
 import java.util.List;
 import java.util.Optional;
 
-import com.ryc.api.v2.club.domain.ClubDetailImage;
-import com.ryc.api.v2.club.domain.ClubSummary;
-import com.ryc.api.v2.club.domain.ClubTag;
+import com.ryc.api.v2.club.domain.vo.ClubDetailImage;
+import com.ryc.api.v2.club.domain.vo.ClubSummary;
+import com.ryc.api.v2.club.domain.vo.ClubTag;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
