+++ server/src/main/java/com/ryc/api/v2/club/presentation/dto/response/ClubUpdateResponse.java
@@ -5,10 +5,10 @@ import java.util.List;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-import com.ryc.api.v2.club.domain.Category;
-import com.ryc.api.v2.club.domain.ClubDetailImage;
-import com.ryc.api.v2.club.domain.ClubSummary;
-import com.ryc.api.v2.club.domain.ClubTag;
+import com.ryc.api.v2.club.domain.enums.Category;
+import com.ryc.api.v2.club.domain.vo.ClubDetailImage;
+import com.ryc.api.v2.club.domain.vo.ClubSummary;
+import com.ryc.api.v2.club.domain.vo.ClubTag;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
