+++ server/src/test/java/com/ryc/api/v2/club/business/ClubAnnouncementFacadeTest.java
@@ -14,7 +14,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.ryc.api.v2.announcement.domain.enums.AnnouncementStatus;
 import com.ryc.api.v2.announcement.service.AnnouncementService;
-import com.ryc.api.v2.club.domain.Club;
+import com.ryc.api.v2.club.domain.vo.Club;
 import com.ryc.api.v2.club.presentation.dto.response.AllClubGetResponse;
 
 @ExtendWith(MockitoExtension.class)
