+++ server/src/test/java/com/ryc/api/v2/club/infra/ClubRepositoryImplTest.java
@@ -1,6 +1,5 @@
 package com.ryc.api.v2.club.infra;
 
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -18,10 +17,10 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.ryc.api.v2.auth.domain.Admin;
-import com.ryc.api.v2.club.domain.Category;
-import com.ryc.api.v2.club.domain.Club;
-import com.ryc.api.v2.club.domain.ClubTag;
-import com.ryc.api.v2.club.domain.Role;
+import com.ryc.api.v2.club.domain.enums.Category;
+import com.ryc.api.v2.club.domain.vo.Club;
+import com.ryc.api.v2.club.domain.vo.ClubTag;
+import com.ryc.api.v2.club.domain.enums.Role;
 import com.ryc.api.v2.club.infra.entity.ClubEntity;
 import com.ryc.api.v2.club.infra.entity.RoleEntity;
 import com.ryc.api.v2.club.infra.jpa.ClubJpaRepository;
