+++ server/src/test/java/com/ryc/api/v2/club/presentation/ClubHttpApiTest.java
@@ -23,8 +23,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ryc.api.v2.club.business.ClubAnnouncementFacade;
 import com.ryc.api.v2.club.business.ClubService;
-import com.ryc.api.v2.club.domain.Category;
-import com.ryc.api.v2.club.domain.ClubTag;
+import com.ryc.api.v2.club.domain.enums.Category;
+import com.ryc.api.v2.club.domain.vo.ClubTag;
 import com.ryc.api.v2.club.presentation.dto.request.ClubCreateRequest;
 import com.ryc.api.v2.club.presentation.dto.request.ClubUpdateRequest;
 import com.ryc.api.v2.club.presentation.dto.response.AllClubGetResponse;
